add insert project endpoint

+++ backend/src/main/java/inf117/projects/repo/entity/Sponsor.java
@@ -1,15 +1,15 @@
 package inf117.projects.repo.entity;
 
 public class Sponsor {
-    private int id;
+    private Long id;
     private String name;
     private String website;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public Sponsor setId(int id) {
+    public Sponsor setId(Long id) {
         this.id = id;
         return this;
     }
